+++ backend/dataall/modules/dataset_sharing/services/share_processors/lf_process_same_account_share.py
@@ -3,6 +3,7 @@ import logging
 from dataall.core.environment.db.environment_models import Environment, EnvironmentGroup
 from dataall.modules.dataset_sharing.common.enums import ShareItemStatus, ShareObjectActions, ShareItemActions
 from dataall.modules.dataset_sharing.db.share_object_models import ShareObject
+from dataall.modules.dataset_sharing.db.share_object_repositories import ShareObjectRepository, ShareItemSM
 from dataall.modules.dataset_sharing.services.share_managers import LFShareManager
 from dataall.modules.datasets_base.db.dataset_models import DatasetTable, Dataset
