+++ backend/dataall/modules/datasets/api/table/__init__.py
@@ -4,7 +4,6 @@ from dataall.modules.datasets.api.table import (
     queries,
     resolvers,
     types,
-    enums,
 )
 
-__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations', 'enums']
+__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations']
