--- backend/dataall/modules/datasets/api/table/enums.py
@@ -1,9 +1,0 @@
-from dataall.base.api.constants import GraphQLEnumMapper
-
-
-class DatasetTablePreviewStatus(GraphQLEnumMapper):
-    QUEUED = 'QUEUED'
-    RUNNING = 'RUNNING'
-    SUCCEEDED = 'SUCCEEDED'
-    FAILED = 'FAILED'
-    CANCELLED = 'CANCELLED'
