+++ backend/dataall/modules/datasets/services/dataset_profiling_service.py
@@ -14,7 +14,7 @@ from dataall.modules.datasets.db.dataset_profiling_repositories import DatasetPr
 from dataall.modules.datasets.db.dataset_table_repositories import DatasetTableRepository
 from dataall.modules.datasets.services.dataset_permissions import PROFILE_DATASET_TABLE, GET_DATASET
 from dataall.modules.datasets_base.db.dataset_repositories import DatasetRepository
-from dataall.modules.datasets_base.constants.enums import ConfidentialityClassification
+from dataall.modules.datasets_base.common.enums import ConfidentialityClassification
 from dataall.modules.datasets_base.db.dataset_models import DatasetProfilingRun, DatasetTable
 from dataall.modules.datasets_base.services.permissions import GET_DATASET_TABLE, PREVIEW_DATASET_TABLE
